handling some comments
